Add a base query to the Cangjie structure
We'll be running SQLite queries, so we need to build those query strings
at some point.

Part of it is static to the Cangjie instance, so we can do it an
initialization time.

This is the first part: it selects what we are interested in, and only
filter on the Cangjie version.