char: Make the constructor return an integer
At the moment, we don't really have a way to report errors to the caller
of our APIs.

We will thus change our functions so that they all return integers
representing the success or errors, as is common in C libraries.

This ports the constructor of the CangjieChar structure, as well as all
the parts of the library and its unit tests which call it.