list: Make the various APIs return integers
At the moment, we don't really have a way to report errors to the caller
of our APIs.

We will thus change our functions so that they all return integers
representing the success or errors, as is common in C libraries.

This ports the various functions related to the CangjieCharList
structure, as well as all the parts of the library and its unit tests
which call it.