Remove one usage of append_string
I don't really like this function, and it's not needed in this case: we
know how much memory we need at most, and we can just concatenate the
two strings.

We don't need any fancy dynamic string management, so let's avoid it, it
will avoid complex code and hard to debug errors.