table: Use 'NA' as the missing code
I'm about to introduce the concept of "short codes" to allow the user to
type e.g '.' and return them things like '。' as a convenience.

However, that means we're going to have '!' as a valid code, whereas so
far it was the code we used to mark that this character had no code.