dbbuilder: Add an index on the version and code
This seems to make each query faster:

* wildcard queries go from taking (in average) 98ms to 76ms
* non-wildcard queries go from taking 79ms to 0.220ms !

That's a huge gain for non-wildcard queries, which touch typist will no
doubt appreciate.

As for wildcard queries, they don't gain so much, but that's not too
much of a problem, given that people using wildcard queries don't expect
speed (either they use the Quick input method, or they forgot the
complete code). And it's not that bad anyway. :)