Fixed typing issue with U+9ec3 and U+9ec4
There is opinion on the character mapping of Chinese character
"黃" (U+9ec3) and "黄" (U+9ec4).

In the previous mapping, "黃" only maps to "tlwc" in Cangjie 5.
That didn't address a character variation introduced by Taiwan
Minster of Education. Detailed description could be found here:

https://github.com/bochecha/ibus-cangjie/issues/28

Since one main source of Cangjie users is Taiwan, it is better to
provide result that compatible to their expectations. Luckily,
libcangjie provides multiple mapping. The change:

The old mapping:

  * tlwc -> 黃 (U+9ec3), 黄 (U+9ec4)

The new mapping:

  * tlwc -> 黃 (U+9ec3), 黄 (U+9ec4)
  * tmwc -> 黃 (U+9ec3), 黄 (U+9ec4)
  * tmlc -> 黃 (U+9ec3)