README: Drop the mention to the AUTHORS file
People are telling me that they don't find the AUTHORS file in the Git
tree, which is normal as it's generated when creating the tarball. I can
understand it is confusing, though.

It's also becoming problematic when we reuse the README as a page of the
website.

All in all, let's remove it and avoid the confusion.