Removed stdlib exit code reference from CLI
References to stdlib exit code (EXIT_SUCCESS, EXIT_FAILURE) was
added in commit aa71faaeb2. In particular, EXIT_FAILURE was added
as a representation of input argument error.

On the other hand, there are number of return statements that
would return the Sqlite3 error code.

Since both EXIT_FAILURE and SQLITE_ERROR are the 1, it is not
possible to distinct the return code between input argument errors
and sqlite errors.

This commit changes all the EXIT_FAILURE to -1 and EXIT_SUCCESS to
1. Also, the reference to stdlib was removed.