cangjie: Use the GLOB operator
We are currently using the LIKE operator to run the SQL queries which
contain a wildcard.

That means we have to transform the wildcard inputted by the user ('*')
into the wildcard accepted by the LIKE operator: '%'.

But SQLite supports the GLOB operator, which uses shell-like globs. That
means its wildcard character is the '*', the same as the one inputted by
the user.

Sounds like a win.