table: Have two frequencies for 氐
The 氐 character has 2 Cangjie codes in our data: 'hpm' and 'hvpi'.

However, depending on the reference sources (dictionaries, CBF's
website, the chinesecj.com forum,...), we can find either only 'hpm',
or both.

That's all a bit confusing...

Doing some tests with Cangjie users, they only knew about the 'hpm'
code, never the 'hvpi' one (although once told about it, they were
usually able to understand how to decompose 氐 in 'hvpi').

That tells me that 'hpm' is absolutely necessary, but not 'hvpi'.

Unfortunately, 'hvpi' is actually causing us trouble, as Quick users
don't expect to see 氐 as the third result when typing 'hi'. In fact,
they don't expect it at all when typing 'hi', only when typing 'hm'.

The way out is to set the frequency to 0 for this character when called
by the 'hvpi' code, so that it appears way low in the list.

That should please Quick users.