Remade README.table.rst to README.table.md
All other README and documentations in this project is in markdown
format. It make sense to change this file to markdown too. Also
removed all number words from the file to make it easier to change.

To adapt the changed filename, also changed the references in
Makefile.am and table.txt