Order the query results
Added "ORDER BY frequency ASC" to the end of query strings used
in cangjie_new_characters and cangjie_new_characters_shortcode.

Since the linked-list is built with an prepend mechanism, the results
will be of reverse order of the query. Which will effectively gives
a "ORDER BY frequency DESC" result.

This commit fixes the default order of results, but it does not provide
a way for user to freely specify sorting logic.

Partially fulfilled enhancement #64.