ci: Parametrize the base image from which we build
We want to run the libcangjie tests on multiple operating systems, for
which we have base build-essential images.

We could use one Dockerfile per OS, and they would all be identical
except for the FROM instruction.

Instead, we can parametrize the FROM instruction to depend on a build
argument, which we pass with the --build-arg option:

* in .travis.yml for the CI
* in the build hook for the Docker Hub (for automated builds)