ci: Don't run the tests in Rawhide for now
We can't build the build-essential:fedora-rawhide base image because the
fedora:rawhide one is too old.

I asked the Fedora release engineering team, and they'll try and push an
update soon.

Once that's done, we should be able to build our base image, and thus to
enable the CI for libcangjie.