wip: data: Secondary codes should have a frequency of 0
Most characters have a single code in a given Cangjie version.

But some characters will have more than one, due to how different
organisms might have applied the Cangjie decomposition method in
different ways.

We try to be as inclusive as possible, so our data includes all those
codes.

However, our focus still is on providing the best possible input
experience to Hong Kong people, and as such we want to give priority to
the codes they are used to.

Those are the codes coming first in our data, and these codes should be
the only ones with a non-zero frequency, so that secondary codes always
are ordered towards the end.

The changes to the data in this commit were made with a script, and
should be entirely reproducible. From a libcangjie clone on the master
branch (commit TODO), get the script and run it:

    $ git clone https://gitlab.freedesktop.org/cangjie/data-migration-tools.git
    $ ./data-migration-tools/fix-secondary-codes-frequency.py

The modifications done by the script to the table should be identical to
the ones in this commit.

Fixes #104