tests: Fix literal strings
This was done a while ago, to make MSVC happier as it doesn't like
non-ascii strings in C code.

(Or at least, that was the case at the time. This was almost 5 years ago
after all…)

    commit 273d280ee0a93e11e1e7651c8697a367f0ddb0ea
    Author: Linquize <linquize@yahoo.com.hk>
    Date:   Thu Oct 17 21:20:36 2013 +0800

        Do not write Chinese literal strings directly

        MSVC does not encoding Chinese literal strings as UTF-8

However, for some reason the backslashes got escaped in this file (and
only in this file), and we missed it on code review.