data: Reorder the codes for 函
We are about to split multiple codes for any given character so that
only the first one has the non-zero frequency, and all additional codes
have a frequency of 0. (see #104)

A prerequisite to that is that the multiple codes are actually ordered
correctly.

On Windows, the reference for Cangjie 3, 函 only has code `nue`.

This means that in Cangjie 3 `nue` should be the primary code for that
character. We keep `neu` but it should have a frequency of 0 so that it
doesn't interfere with the expected ordering.