data: Split the short codes to their own lines
Currently, the frequency is associated to a character and all its codes,
except for a single exception.

We are about to split multiple codes for any given character so that the
frequency becomes instead associated to a code for a given character.
(see #104)

This commit splits the short codes to their own lines, so that they can
have their own frequencies, separated from the frequencies associated to
each Cangjie 3 and 5 codes.